Refactor : 알림 경로 수정
